Integer -> Long 타입 변경, email주문조회 추가,

+++ backend/src/main/java/com/ll/coffee/order/OrderDto.java
@@ -10,13 +10,13 @@ import java.util.Map;
  */
 @Getter
 public class OrderDto {
-    private int id;
+    private long id;
     private String email;
     private String address;
     private int postalCode;
-    private Map<Integer, Integer> orders;
+    private Map<Long, Integer> orders;
 
-    public OrderDto(Order order, Map<Integer, Integer> orders) {
+    public OrderDto(Order order, Map<Long, Integer> orders) {
         this.id = order.getId();
         this.email = order.getEmail();
         this.address = order.getAddress();
